requirements.txt
apache-airflow>=2.7.1
requests>=2.31.0
overpy>=0.6
pandas>=1.5.0,<2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
SQLAlchemy>=1.4.36,<2.0.0
openai>=1.3.0
geopandas>=0.12.0
shapely>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Linting
flake8>=6.1.0
black>=23.7.0
isort>=5.12.0

# Additional dependencies
numpy>=1.24.0
apache-airflow-providers-postgres>=5.7.1 